requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
iniconfig==2.0.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
packaging==23.2
pluggy==1.3.0
pytest==7.4.3
pytest-check==2.2.2
referencing==0.30.2
requests==2.31.0
rpds-py==0.12.0
tenacity==8.2.3
urllib3==2.1.0
